Move autocomplete interaction to team.ts
