+++ src/interactions/team.ts
@@ -1,4 +1,5 @@
 import {
+    AutocompleteInteraction,
     CommandInteraction,
     EmbedBuilder,
     SlashCommandBuilder,
@@ -57,5 +58,27 @@ Speed: \`${team.speed}\` Megapixelsteps per second`,
         });
 
         await interaction.followUp({ embeds: [embed] });
+    },
+    async autocompleteHandler(interaction: AutocompleteInteraction) {
+        const query = interaction.options.getFocused(true).value.toLowerCase();
+
+        const { data: teams } = await API.getTeams();
+
+        if (teams.length === 0) return interaction.respond([]);
+
+        const selectedTeams = teams.filter(
+            (t) =>
+                t.name?.toLowerCase().includes(query) ||
+                t.id?.toLowerCase().includes(query)
+        );
+
+        return interaction.respond(
+            selectedTeams
+                .map((t) => ({
+                    name: t.name || "",
+                    value: t.id || "",
+                }))
+                .slice(0, 25)
+        );
     },
 };
